svn merge -r207499:207689 svn+ssh://gcc.gnu.org/svn/gcc/trunk
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/redhat@207690 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc-4_9-branch/gcc/testsuite/gcc.dg/vect/pr60012.c
@@ -1,4 +1,5 @@
 /* { dg-do compile } */
+/* { dg-require-effective-target vect_int } */
 /* { dg-additional-options "--param vect-max-version-for-alias-checks=0" } */
 
 typedef struct
@@ -8,14 +9,14 @@ typedef struct
 } complex16_t;
 
 void
-libvector_AccSquareNorm_ref (unsigned long long  *acc,
+libvector_AccSquareNorm_ref (unsigned int *acc,
 			     const complex16_t *x, unsigned len)
 {
   unsigned i;
   for (i = 0; i < len; i++)
-    acc[i] += ((unsigned long long)((int)x[i].real * x[i].real))
-	+ ((unsigned long long)((int)x[i].imag * x[i].imag));
+    acc[i] += ((unsigned int)((int)x[i].real * x[i].real))
+	+ ((unsigned int)((int)x[i].imag * x[i].imag));
 }
 
-/* { dg-final { scan-tree-dump "LOOP VECTORIZED" "vect" } } */
+/* { dg-final { scan-tree-dump "LOOP VECTORIZED" "vect" { target { vect_extract_even_odd } } } } */
 /* { dg-final { cleanup-tree-dump "vect" } } */
